requirements.txt
aioredis==2.0.1
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
async-timeout==4.0.2
bcrypt==3.2.0
brotlipy==0.7.0
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
cryptography==36.0.0
ecdsa==0.17.0
fastapi==0.75.1
future==0.18.2
gunicorn==20.1.0
h11==0.13.0
httptools==0.3.0
idna==3.3
orjson==3.6.7
passlib==1.7.4
pip==21.2.4
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.10.1
pydantic==1.9.0
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.27.1
rsa==4.7.2
setuptools==58.0.4
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.1
wheel==0.37.1
